digest/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11
Jinja2==3.1.4
thefuzz==0.22.1
python-Levenshtein==0.25.1
python-dotenv==1.0.1
